SineWaves.jl

SineWaves.SineWave — Type

SineWave(frequency, samplerate)

Structure that has fields corresponding to C library structure. It is initialized by calling C init function from a private constructor.

Base.fill! — Function

fill!(buffer, sinewave)

Fills buffer with next sinewave generator samples by calling fill function from the C library.

SineWaves.spectrum — Function

spectrum(buffer)

Gets spectrum of buffer argument. It calls spectrum function from the C library which allocates new array with malloc. This memory is reclaimed later by Julia's garbage collector thanks to invocation of unsafe_wrap method with flag own equal to true.